pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "axle"
version = "0.0.1"
description = "A composable toolset for agentic video workflows"
readme = "README.md"
authors = [{ name = "Dany Boza" }]
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12.0",
    "libsql>=0.0.1",
    "google-genai>=1.0.0",
    "deepgram-sdk>=3.0.0",
    "boto3>=1.34.0",
    "pydantic>=2.7.0",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
axle = "axle.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["axle*"]
